requirements.txt
# Telegram Bot
aiogram>=3.0
python-dotenv>=1.0
pydantic-settings>=2.1.0
pydantic>=2.5.0

# AI/LLM
google-generativeai>=0.3
tiktoken>=0.7.0
tenacity>=8.2.3

# Caching & Storage
redis>=5.0
msgpack>=1.0.7

# Speech Recognition
SpeechRecognition>=3.10
pydub>=0.25
transformers>=4.21.0
torch --index-url https://download.pytorch.org/whl/cu130
torchvision --index-url https://download.pytorch.org/whl/cu130
soundfile>=0.12.0

# File Operations
aiofiles>=23.0

# Semantic Search & Embeddings
sentence-transformers>=2.2.0
numpy>=1.21.0

# Progress Bars
tqdm>=4.65.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0